requirements.txt
requests>=2.25.0
beautifulsoup4>=4.9.0
selenium>=3.141.0
pandas>=1.1.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
streamlit>=1.0.0
plotly>=5.0.0
